+++ src/main/resources/edu/wpi/N/css/adminTables.css
@@ -1,47 +1,46 @@
 .table-view{
-    -fx-background-color: #7fd9ec;
+    -fx-background-color: #6C5C7F;
+    -fx-background-radius: 30px;
 }
 
 .table-view:focused{
     -fx-background-color: #7fd9ec;
 }
 
+.table-view .column-header-background {
+    -fx-background-color: #7E9AB6;
+    -fx-background-radius: 30;
+    -fx-background-insets: 0 0 10 0;
+    -fx-padding: 0 0 10 0;
 
-.table-view .column-header-background{
-    -fx-background-color: #002186;
+    /*
+    -fx-effect: dropshadow(two-pass-box, rgb(10, 101, 107), 15, 0, 0, 0);
+     */
 }
 
-.table-view .column-header-background .label{
+.table-view .column-header, .table-view .filler {
+    -fx-size: 30;
+    -fx-border-width: 0 0 10 0;
     -fx-background-color: transparent;
-    -fx-text-fill: white;
+    -fx-background-radius: 30;
 }
 
-.table-view .column-header {
-    -fx-background-color: #002186;
+.table-view .column-header .label {
+    -fx-text-fill: #E6EBF2;
 }
 
-.table-view .table-cell{
-    -fx-text-fill: black;
+.table-view .table-cell {
+    -fx-text-fill: #E6EBF2;
 }
 
-.table-row-cell{
-    -fx-background-color: -fx-table-cell-border-color, #b1e8f4;
-    -fx-background-insets: 0, 0 0 1 0;
-    -fx-padding: 0.0em; /* 0 */
-}
-
-.table-row-cell:odd{
-    -fx-background-color: -fx-table-cell-border-color, #7fd9ec;
-    -fx-background-insets: 0, 0 0 1 0;
-    -fx-padding: 0.0em; /* 0 */
+.table-row-cell {
+    -fx-background-color: #6C5C7F;
 }
 
 .table-row-cell:selected {
-    -fx-background-color: #005797;
-    -fx-background-insets: 0;
-    -fx-background-radius: 1;
+    -fx-background-color: #7E9AB6;
 }
-
+/*
 .choice-box .context-menu {
     -fx-background-color: white;
     -fx-text-fill: black;
@@ -53,3 +52,4 @@
 }
 
 .choice-box > .label { -fx-text-fill: white; }
+ */
