+++ src/main/resources/edu/wpi/N/views/admin/viewRequest.fxml
@@ -2,6 +2,7 @@
 
 <?import com.jfoenix.controls.JFXButton?>
 <?import com.jfoenix.controls.JFXCheckBox?>
+<?import com.jfoenix.controls.JFXTreeTableView?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ChoiceBox?>
 <?import javafx.scene.control.Label?>
@@ -11,19 +12,19 @@
 
 <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="970.0" prefWidth="1720.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.wpi.N.views.admin.ViewRequestController">
    <children>
-      <Label layoutX="901.0" layoutY="242.0" text="Service Requests">
+      <Label layoutX="729.0" layoutY="776.0" text="Service Requests">
          <font>
             <Font name="Calibri" size="32.0" />
          </font>
       </Label>
-      <TableView fx:id="tb_RequestTable" layoutX="611.0" layoutY="317.0" maxWidth="-Infinity" minWidth="-Infinity" prefHeight="429.0" prefWidth="955.0" stylesheets="@../../css/adminTables.css" AnchorPane.leftAnchor="611.0" AnchorPane.rightAnchor="354.0" />
+      <TableView fx:id="tb_RequestTable" layoutX="611.0" layoutY="317.0" maxWidth="-Infinity" minWidth="-Infinity" prefHeight="429.0" prefWidth="955.0" stylesheets="@../../css/adminTables.css" visible="false" AnchorPane.leftAnchor="611.0" AnchorPane.rightAnchor="354.0" />
       <Button fx:id="btn_editMap" disable="true" mnemonicParsing="false" prefHeight="50.0" prefWidth="190.0" text="Edit Map" visible="false">
          <font>
             <Font name="Calibri" size="24.0" />
          </font>
       </Button>
-      <ChoiceBox fx:id="cb_Employee" layoutX="424.0" layoutY="424.0" prefWidth="150.0" />
-      <TableView fx:id="tb_languages" layoutX="424.0" layoutY="459.0" prefHeight="202.0" prefWidth="150.0" stylesheets="@../../css/adminTables.css" />
+      <ChoiceBox fx:id="cb_Employee" layoutX="1499.0" layoutY="292.0" prefWidth="150.0" />
+      <TableView fx:id="tb_languages" layoutX="1438.0" layoutY="478.0" prefHeight="202.0" prefWidth="150.0" stylesheets="@../../css/adminTables.css" visible="false" />
       <JFXButton layoutX="1190.0" layoutY="765.0" onMouseClicked="#assignPressed" text="Assign">
          <font>
             <Font name="Calibri" size="18.0" />
@@ -44,5 +45,7 @@
             <Font name="Calibri" size="12.0" />
          </font>
       </JFXCheckBox>
+      <JFXTreeTableView fx:id="trtbl_requests" layoutX="472.0" layoutY="140.0" prefHeight="583.0" prefWidth="955.0" sortMode="ONLY_FIRST_LEVEL" style="-fx-fixed-cell-size: 150px;" />
+      <JFXTreeTableView fx:id="trtbl_languages" layoutX="1450.0" layoutY="322.0" />
    </children>
 </AnchorPane>
